+++ tests/testthat/test_Learner.R
@@ -327,16 +327,16 @@ test_that("Models can be replaced", {
 
 test_that("test/holdout task's backend is removed", {
   learner = lrn("regr.rpart")
-  task = tsk("mtcars")$partition(1:10, "test")
-  task = tsk("mtcars")$partition(11:20, "holdout")
+  task = tsk("mtcars")$divide(1:10, "test")
+  task = tsk("mtcars")$divide(11:20, "holdout")
   learner$train(task)
   expect_true(is.null(learner$state$train_task$test_task$backend))
   expect_true(is.null(learner$state$train_task$holdout_task$backend))
 })
 test_that("test/holdout task's backend is removed", {
   learner = lrn("regr.rpart")
-  task = tsk("mtcars")$partition(1:10, "test")
-  task = tsk("mtcars")$partition(11:20, "holdout")
+  task = tsk("mtcars")$divide(1:10, "test")
+  task = tsk("mtcars")$divide(11:20, "holdout")
   learner$train(task)
   expect_true(is.null(learner$state$train_task$test_task$backend))
   expect_true(is.null(learner$state$train_task$holdout_task$backend))