+++ tests/testthat/test_Task.R
@@ -380,7 +380,7 @@ test_that("Task$set_row_roles", {
   task$set_row_roles(1:10, add_to = "use")
   expect_true(all(1:10 %in% task$row_ids))
 
-  task$partition(1:10, "holdout")
+  task$divide(1:10, "holdout")
   expect_true(all(1:10 %nin% task$row_ids))
   expect_true(all(1:10 %in% task$holdout_task$row_ids))
 })
@@ -511,7 +511,7 @@ test_that("Roles get printed (#877)", {
 
 test_that("test task is cloned", {
   task = tsk("iris")
-  task$partition(c(1:10, 51:60, 101:110), "test")
+  task$divide(c(1:10, 51:60, 101:110), "test")
   task2 = task$clone(deep = TRUE)
   expect_false(identical(task$test_task, task2$test_task))
   expect_equal(task$test_task, task2$test_task)
@@ -519,7 +519,7 @@ test_that("test task is cloned", {
 
 test_that("holdout task is cloned", {
   task = tsk("iris")
-  task$partition(c(1:10, 51:60, 101:110), "holdout")
+  task$divide(c(1:10, 51:60, 101:110), "holdout")
   task2 = task$clone(deep = TRUE)
   expect_false(identical(task$holdout_task, task2$holdout_task))
   expect_equal(task$holdout_task, task2$holdout_task)
@@ -533,75 +533,75 @@ test_that("task cannot be its own test / holdout task", {
 
 test_that("test task cannot have a test or holdout task", {
   task = tsk("iris")
-  expect_error({task$test_task = task$clone(deep = TRUE)$partition(1, "test") }, "remove its test/holdout")
-  expect_error({task$test_task = task$clone(deep = TRUE)$partition(1, "holdout") }, "remove its test/holdout")
-  expect_error({task$holdout_task = task$clone(deep = TRUE)$partition(1, "test") }, "remove its test/holdout")
-  expect_error({task$holdout_task = task$clone(deep = TRUE)$partition(1, "holdout") }, "remove its test/holdout")
+  expect_error({task$test_task = task$clone(deep = TRUE)$divide(1, "test") }, "remove its test/holdout")
+  expect_error({task$test_task = task$clone(deep = TRUE)$divide(1, "holdout") }, "remove its test/holdout")
+  expect_error({task$holdout_task = task$clone(deep = TRUE)$divide(1, "test") }, "remove its test/holdout")
+  expect_error({task$holdout_task = task$clone(deep = TRUE)$divide(1, "holdout") }, "remove its test/holdout")
 })
 
 test_that("test task changes a task's hash", {
   task = tsk("iris")
   h1 = task$hash
-  task$partition(1:10, "test", remove = FALSE)
+  task$divide(1:10, "test", remove = FALSE)
   h2 = task$hash
   expect_false(h1 == h2)
-  task$partition(1:11, "test", remove = FALSE)
+  task$divide(1:11, "test", remove = FALSE)
   expect_false(h2 == task$hash)
 })
 
 test_that("holdout task changes a task's hash", {
   task = tsk("iris")
   h1 = task$hash
-  task$partition(1:10, "holdout", remove = FALSE)
+  task$divide(1:10, "holdout", remove = FALSE)
   h2 = task$hash
   expect_false(h1 == h2)
-  task$partition(1:11, "holdout", remove = FALSE)
+  task$divide(1:11, "holdout", remove = FALSE)
   expect_false(h2 == task$hash)
 })
 
 test_that("can NULL test/holdout task", {
   task = tsk("iris")
-  task$partition(1, "test")
+  task$divide(1, "test")
   task$test_task = NULL
   expect_equal(length(task$row_ids), 149)
 
   task = tsk("iris")
-  task$partition(1, "holdout")
+  task$divide(1, "holdout")
   task$holdout_task = NULL
   expect_equal(length(task$row_ids), 149)
 })
 
 test_that("can call partition twice", {
   task = tsk("iris")
-  task$partition(1:10, "test")
-  expect_task(task$partition(1:10, "test"))
+  task$divide(1:10, "test")
+  expect_task(task$divide(1:10, "test"))
 
   task = tsk("iris")
-  task$partition(1:10, "holdout")
-  expect_task(task$partition(1:10, "holdout"))
+  task$divide(1:10, "holdout")
+  expect_task(task$divide(1:10, "holdout"))
   task = tsk("iris")
-  task$partition(1:10, "holdout")
-  expect_task(task$partition(1:10, "test"))
+  task$divide(1:10, "holdout")
+  expect_task(task$divide(1:10, "test"))
 })
 
 test_that("holdout task cannot have a test or holdout task", {
   task = tsk("iris")
   task2 = tsk("iris")
-  task2$partition(1, "test")
+  task2$divide(1, "test")
   expect_error({task$test_task = task2}, "remove its")
   expect_error({task$holdout_task = task2}, "remove its")
   task2$test_task = NULL
-  task2$partition(1, "holdout")
+  task2$divide(1, "holdout")
   expect_error({task$test_task = task2}, "remove its")
   expect_error({task$holdout_task = task2}, "remove its")
 })
 
 test_that("test_task is printed", {
   task = tsk("iris")
-  task$partition(c(1:10, 51:60, 101:110), "test")
+  task$divide(c(1:10, 51:60, 101:110), "test")
   out = capture_output(print(task))
   expect_true(grepl(pattern = "* Test Task: (30x5)", fixed = TRUE, x = out))
-  task$partition(c(1:5, 51:60, 101:110), "holdout")
+  task$divide(c(1:5, 51:60, 101:110), "holdout")
   out2 = capture_output(print(task))
   expect_true(grepl(pattern = "* Holdout Task: (25x5)", fixed = TRUE, x = out2))
 })
@@ -611,7 +611,7 @@ test_that("task hashes during resample", {
   task = orig$clone(deep = TRUE)
   resampling = rsmp("holdout")
   resampling$instantiate(task)
-  task$partition(resampling$test_set(1), "test")
+  task$divide(resampling$test_set(1), "test")
   task$hash
   learner = lrn("classif.debug", uses_test_task = TRUE)
   expect_equal(resampling_task_hashes(task, resampling, learner), task$hash)